fix mocha for functional tests (#4907)
* fix mocha for functional tests

* wrap fs.unlink in try/catch since mocha executes the `afterEach` hook after update

* use existsSync

+++ test/functional/fixtures/run-options/stop-on-first-fail/test.js
@@ -13,7 +13,12 @@ const getTestRunCount = () => {
 
 describe('Stop test task on first failed test', () => {
     afterEach(() => {
-        fs.unlinkSync(TEST_RUN_COUNT_FILENAME);
+        // NOTE: after mocha is updated to `^7.1.1` the `afterEach` hook is called if the test is skipped/pending
+        // before the update the `afterEach` hook was not called.
+        // When we run the test not in `chrome` the file will not exist, since we have the { only: 'chrome' } option,
+        // so we cannot unlink it.
+        if (fs.existsSync(TEST_RUN_COUNT_FILENAME))
+            fs.unlinkSync(TEST_RUN_COUNT_FILENAME);
     });
 
     it('Basic', () => {
